fix broken CommandFailedExceptions
In commit 725bc4755bad4f253c49298ea0102fdbaba0e077, a "default" line was
inserted into the results checking switch statement that causes
execute() to throw CommandFailedException EVERY SINGLE TIME a command is
exeucted, even if the command was successful.  This is clearly wrong,
and results in every single command logging a message like

    EPP command failed due to an internal server error.

Even on "successful" responses.

Fixed by moving the default to the end with no action, which is the way
it should have been in the first place.