RPE-1269: Update Toolkit to generate XML with namespace prefixes
This is a request from brandmate team.

The key change is from the `Command.toXML()`, a overwriten method would
accept a `XmlOutputConfig`, via which the caller could specify whether
they want the namespace prefixes to be output.

Noteworth - the prefix is coming from the `NamespaceContextImpl`, so at
this stage it does not allow user to freely specify the prefix. Actually
from the design perspective there is no obvious reason why we would want
to allow caller to freely pass in any prefix. As a toolkit it would be
better to output the prefixes in a consistent manner.

A few test cases have been added to verify the changes.