Decommission the build env for openjdk7 and oraclejdk7
Travis CI has problem on these two building environments,

1. In oraclejdk7 env, './gradlew assemble' failed due to "JAVA_HOME is
    set to an invalid directory: /usr/lib/jvm/java-7-oracle"
2. In openjdk7 env, './gradlew assemble' failed due to
    "java.security.NoSuchProviderException: no such provider: SunEC".

The first issue is because Trusty no longer supports Oracle Java 7,
according to https://github.com/travis-ci/travis-ci/issues/7884.

The second issue is a problem between JVMs and Gradle Inc. servers
(`services.gradle.org` and `plugins.gradle.org`), that can only be
solved by JVM or server configuration changes. The first approach
requires the manipulation of java security providers. The second option
is unlikely to happen because it decreases security, and more
importantly the servers are managed by a third party.
Refer to: https://github.com/gradle/gradle/issues/2421

In the gradle script we actually already force both source compatibility
and target compatibility to be 1.6, which means the build will fail if
the source code is not 1.6 compatible, also it guarantees the generated
class files would be able to run on Java 6+.

That being said, we don't really need to run the builds on JDK7 env. So
this change is to decommission the two build env in Travis CI.