Update readme for TLS protocols, clean up SSLSocket setup
With the recently merged code changes, the toolkit is able to use
different TLS protocols. So the corresponding section in README gets
updated.

It's also noticed during the code review that, `SSLSocket` actually
honours the `SSLContext.getInstance(protocol)` so it probably would be
better off removing the call to `SSLSocket.setEnabledProtocols()` as we
don't really need to override those enabled-protocols.