RPE-1657: Fix logging to avoid false alarm
The logging of EPP session rate limit exceeded should not be in the for
loop as it is still trying to find an available session with minimum
count.

For example, let's say there are 3 sessions available in the pool,

+ Session A (count = 2)
+ Session B (count = 4)
+ Session C (count = 0)

While it's checking the session B, the logging will be triggered, but
actually it will find session C as a best availabe result.